Full regression sweep + memory audit: fix a real ~1GB-per-call leak
Requested full check: re-ran the complete verification suite fresh
(702 assertions across all 11 batches), loaded all 39 engine modules
together in one process (catches cross-module issues no per-batch
isolated test run would surface), syntax-checked every file, and
confirmed a clean git tree. Zero bugs found in this pass beyond the
memory issues below.

Requested memory audit: user asked for the new engine/ library (not
the deployed render pipeline, which already has its own carefully-
tuned, battle-tested memory settings - see longVideoOrchestrator.js's
extensive comments on why a too-tight V8 heap cap previously caused
real production chunk-timeout failures) to stay under 150MB per call.

Two real fixes, found via direct process.memoryUsage() measurement at
realistic resolutions, not assumed:

1. blurEffects.js's convolve1D/convolveDirectional/radialBlur used
   Float64Array for their width*height*4-element intermediate buffers.
   Measured: a single gaussianBlur call at 1080x1920 cost ~125MB RSS.
   Float32Array has ~7 decimal digits of precision - vastly more than
   an 8-bit color channel needs before the final round-to-byte step -
   so switching halves the dominant cost for free, with zero output
   change (full batch 9 suite still passes). Also applied to
   motionTracking.js's toGray() (a full-frame-sized buffer). Result at
   1080x1920: ~70MB, safely under budget.

2. FAR more significant: layer3d.js's warpTriangle (batch 7's exact
   per-triangle affine warp - used by every 3D layer, the card3DFlip
   transition, AND puppetTool.js's mesh deformation) called
   ctx.drawImage(sourceCanvas, 0, 0) on the FULL source canvas on
   EVERY triangle, relying on a clip to discard everything outside
   that one small triangle. Measured: warpImageToQuad's default 8x8
   grid (128 triangles) cost over 1GB RSS at 1080x1920 - a single call.

   First fix attempt (crop the source via drawImage's own 9-argument
   sub-rectangle form before warping) measurably did NOT help -
   isolated directly why: @napi-rs/canvas's drawImage() pays a real
   cost proportional to the SOURCE canvas's own full size on every
   call, REGARDLESS of destination size or how small a sub-rectangle
   is requested (confirmed: a 52x52 crop FROM a 1080x1920 source still
   cost ~8MB per call, same as drawing the whole thing) - the crop
   step itself was still touching the huge source.

   Real fix: cache each source canvas's full pixel data via ONE
   getImageData call (keyed by the canvas object in a WeakMap, so it
   self-cleans the moment nothing else references that canvas - no
   explicit cleanup, no unbounded growth), then build each triangle's
   small crop via plain typed-array byte copying - no drawImage call
   on the large source at any point, ever again. getImageData does not
   have the same repeat-per-call cost; it's a genuine one-time,
   size-proportional read. Verified mathematically identical output
   (a crop is lossless, the affine transform is re-solved against the
   crop's own local origin) via the full batch 7/8/10/11 suites (289
   assertions covering this exact code path) plus a visual re-render
   of both the batch 10 (puppet mesh) and batch 11 (3D card flip)
   integration demos, pixel-identical to before the fix.

   Result at 1080x1920: 1GB+ -> 82.9MB (a ~12x reduction). At the
   actual current production render resolution (540x960): 24.6MB.

After both fixes: full 702-assertion suite still passes with zero
regressions. Honest ceiling: an extreme 4K-resolution single call
still exceeds 150MB (gaussianBlur ~317MB, warpImageToQuad ~321MB at
2160x3840) - the fundamental buffer/allocation size at that pixel
count is simply large regardless of these fixes, and getting a hard
guarantee at any resolution would need a further architectural change
(tiled/streaming processing) not undertaken here since it isn't
needed for any currently-relevant use (production renders at 540x960;
even a generous 1080x1920 full-HD ceiling is now comfortably safe).

+++ backend/engine/blurEffects.js
@@ -38,9 +38,26 @@ function buildBoxKernel(radius) {
   return { kernel, radius };
 }
 
-/** One axis-aligned 1D convolution pass. Edge handling clamps to the nearest valid pixel (the standard "extend" boundary condition) - the visually natural choice for a blur, avoiding the artificial dark fringe a zero-padded boundary would introduce. */
+/**
+ * One axis-aligned 1D convolution pass. Edge handling clamps to the
+ * nearest valid pixel (the standard "extend" boundary condition) - the
+ * visually natural choice for a blur, avoiding the artificial dark
+ * fringe a zero-padded boundary would introduce.
+ *
+ * The output buffer is Float32Array, not Float64Array - measured
+ * directly before making this change: at 1080x1920 (a real, common
+ * vertical-video resolution), a single gaussianBlur call's RSS grew by
+ * ~125MB, and at 2160x3840 (4K) by over 500MB, dominated by exactly
+ * these width*height*4-element intermediate buffers at 8 bytes/element.
+ * 32-bit float has ~7 decimal digits of precision - vastly more than
+ * an 8-bit (0-255) color channel needs before the final round-to-byte
+ * step discards any difference anyway - so this halves the dominant
+ * memory cost with zero measurable effect on output (verified: the
+ * full batch 9 test suite, including exact energy-preservation and
+ * boundary-value assertions, still passes unchanged).
+ */
 function convolve1D(data, width, height, kernel, radius, horizontal) {
-  const out = new Float64Array(data.length);
+  const out = new Float32Array(data.length);
   for (let y = 0; y < height; y++) {
     for (let x = 0; x < width; x++) {
       let r = 0, g = 0, b = 0, a = 0;
@@ -87,7 +104,7 @@ function boxBlur(imageData, { radius = 8, iterations = 1 } = {}) {
 /** One 1D convolution pass along an ARBITRARY angle (not just horizontal/vertical) - samples along the line through each pixel at the given direction, rounded to the nearest source pixel (a reasonable simplification for a many-tap blur kernel - averaging several nearest-pixel samples along a line washes out individual rounding error, unlike a single bilinear sample would need to avoid visible stepping). */
 function convolveDirectional(data, width, height, kernel, radius, angleRad) {
   const dx = Math.cos(angleRad), dy = Math.sin(angleRad);
-  const out = new Float64Array(data.length);
+  const out = new Float32Array(data.length); // Float32, not Float64 - see convolve1D's doc comment for the measured reason
   for (let y = 0; y < height; y++) {
     for (let x = 0; x < width; x++) {
       let r = 0, g = 0, b = 0, a = 0;
@@ -131,7 +148,7 @@ function radialBlur(imageData, { amount = 10, center = null, mode = 'zoom', samp
   const { width, height, data } = imageData;
   const c = center || [width / 2, height / 2];
   const src = new Uint8ClampedArray(data);
-  const out = new Float64Array(data.length);
+  const out = new Float32Array(data.length); // Float32, not Float64 - see convolve1D's doc comment for the measured reason
 
   for (let y = 0; y < height; y++) {
     for (let x = 0; x < width; x++) {
